test: 웰컴페이지 테스트 클래스 이동
