+++ README.md
@@ -32,4 +32,4 @@
 ### 3단계 요구사항
 
 - [x] 날짜와 테마를 기반으로 예약 가능한 시각과 예약 불가능한 시각을 응답할 수 있다.
-- [ ] `/`요청 시 인기 테마 페이지를 응답한다. 
+- [x] `/`요청 시 인기 테마 페이지를 응답한다.
