+++ src/test/java/roomescape/controller/view/ClientViewControllerTest.java
@@ -1,0 +1,23 @@
+package roomescape.controller.view;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class ClientViewControllerTest {
+
+    @DisplayName("웰컴페이지 요청 시 응답할 수 있다(200 OK)")
+    @Test
+    void should_response_200_when_request_welcome_page() {
+        RestAssured.given().log().all()
+                .when().get("/")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
